Create build and install mods
